bugfix submit workflow command with building AIGraph (#254)
